tag_bot/requirements.txt
praw==6.5.1
prawcore==1.0.1
python-dotenv==0.13.0
